added serving static files to backend, made small change to show it on frontend, updated heroku procfile and added scripts
