+++ Procfile
@@ -1,0 +1,1 @@
+web: NODE_ENV=production node backend/src/server.js
