+++ frontend/public/index.html
@@ -36,5 +36,6 @@
       To begin the development, run `npm start` or `yarn start`.
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
+    <p>World's Greatest Event-app</p>
   </body>
 </html>
